sustainableGardenBackend/requirements.txt
Django == 3.2
djangorestframework == 3.12.4
django-crontab ==0.7.1
pyserial == 3.4
requests == 2.21.0
psycopg2 == 2.9.1
django-cors-headers